int32 to double + round

+++ WPF App/WPF App/Model/Improvement.cs
@@ -10,11 +10,11 @@ namespace WPF_App.Model
     {
         private int _numberOfUpgrades;
         private int _startingPrice;
-        private int _currentPrice;
+        private double _currentPrice;
 
         public int NumberOfUpgrades { get => _numberOfUpgrades; set => _numberOfUpgrades = value; }
         public int StartingPrice { get => _startingPrice; set => _startingPrice = value; }
-        public int CurrentPrice { get => _currentPrice; set => _currentPrice = value; }
+        public double CurrentPrice { get => _currentPrice; set => _currentPrice = value; }
 
         // METHODS
         public abstract void Upgrade();
